+++ internal/topic/topicreaderinternal/stream_reader_impl.go
@@ -23,9 +23,11 @@ import (
 )
 
 var (
-	errPartitionStopped              = xerrors.Wrap(errors.New("ydb: pq partition stopped"))
-	errCommitSessionFromOtherReader  = xerrors.Wrap(errors.New("ydb: commit with session from other reader"))
-	errCommitWithNilPartitionSession = xerrors.Wrap(errors.New("ydb: commit with nil partition session"))
+	errPartitionStopped                                = xerrors.Wrap(errors.New("ydb: pq partition stopped"))
+	errCommitSessionFromOtherReader                    = xerrors.Wrap(errors.New("ydb: commit with session from other reader")) //nolint:lll
+	errCommitWithNilPartitionSession                   = xerrors.Wrap(errors.New("ydb: commit with nil partition session"))
+	errInternalUnknownCommitMode                       = xerrors.Wrap(errors.New("ydb: unexpected commit mode"))
+	errInternalSaveAutoCommitBeforeCommitPreviousRange = xerrors.Wrap(errors.New("ydb: internal state error - save new range for autocommit before commit previous")) //nolint:lll
 )
 
 type partitionSessionID = rawtopicreader.PartitionSessionID
@@ -37,6 +39,7 @@ type topicStreamReaderImpl struct {
 
 	freeBytes                 chan int
 	atomicRestBufferSizeBytes int64
+	lastReadRange             commitRange
 	sessionController         partitionSessionStorage
 	backgroundWorkers         background.Worker
 
@@ -75,7 +78,7 @@ func newTopicStreamReaderConfig() topicStreamReaderConfig {
 		BufferSizeProtoBytes:  1024 * 1024,
 		Cred:                  credentials.NewAnonymousCredentials(),
 		CredUpdateInterval:    time.Hour,
-		CommitMode:            CommitModeAsync,
+		CommitMode:            CommitModeAutoCommit,
 		CommitterBatchTimeLag: time.Second,
 		Decoders:              newDecoderMap(),
 	}
@@ -171,10 +174,45 @@ func (r *topicStreamReaderImpl) ReadMessageBatch(
 	defer func() {
 		if err == nil {
 			r.freeBufferFromMessages(batch)
+
+			if r.cfg.CommitMode == CommitModeAutoCommit {
+				err = r.autoCommitStore(batch.commitRange)
+			}
 		}
 	}()
 
+	if r.cfg.CommitMode == CommitModeAutoCommit {
+		if err = r.autoCommit(ctx); err != nil {
+			return nil, err
+		}
+	}
+
 	return r.consumeMessagesUntilBatch(ctx, opts)
+}
+
+func (r *topicStreamReaderImpl) autoCommit(ctx context.Context) error {
+	if r.lastReadRange.session() == nil || r.lastReadRange.session().Context().Err() != nil {
+		return nil
+	}
+
+	err := r.commit(ctx, r.lastReadRange)
+	if errors.Is(err, errPartitionStopped) {
+		err = nil
+	}
+
+	if err != nil {
+		return err
+	}
+	return r.autoCommitStore(commitRange{})
+}
+
+func (r *topicStreamReaderImpl) autoCommitStore(lastReadRange commitRange) error {
+	if r.lastReadRange.session() != nil && lastReadRange.session() != nil {
+		return errInternalSaveAutoCommitBeforeCommitPreviousRange
+	}
+
+	r.lastReadRange = lastReadRange
+	return nil
 }
 
 func (r *topicStreamReaderImpl) consumeMessagesUntilBatch(
@@ -298,6 +336,19 @@ func (r *topicStreamReaderImpl) onPartitionSessionStatusResponseFromBuffer(
 }
 
 func (r *topicStreamReaderImpl) Commit(ctx context.Context, commitRange commitRange) error {
+	switch r.cfg.CommitMode {
+	case CommitModeAutoCommit, CommitModeNone:
+		return ErrCommitDisabled
+	case CommitModeAsync, CommitModeSync:
+		// pass
+	default:
+		return xerrors.WithStackTrace(errInternalUnknownCommitMode)
+	}
+
+	return r.commit(ctx, commitRange)
+}
+
+func (r *topicStreamReaderImpl) commit(ctx context.Context, commitRange commitRange) error {
 	if err := r.checkCommitRange(commitRange); err != nil {
 		return err
 	}